pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "dro_circuit"
version = "0.0.1"
description = "DRO over corruptions for faithful circuit discovery in transformers."
authors = [
  { name = "Jiahao Zhang" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "torch>=2.1",
    "transformer-lens>=2.15.0",
    "numpy",
    "matplotlib>=3.10",
    "pyyaml",
    "tqdm",
    "pytest",
    "python-dotenv",
    "ruff",
]
requires-python = ">=3.10"


[tool.ruff]
line-length = 99
src = ["dro_circuit"]
include = ["pyproject.toml", "dro_circuit/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["dro_circuit"]
force-sort-within-sections = true

